Update pair_programming.md
Olivier: Create a RoR webservice that grants authorisation without the need to log in. A bit like doodle where you can create and edit a doodle as the organiser without logging in into the service. As a participant you can join and edit your participation to the doodle without logging in either.